display_setup/pyproject.toml
[project]
name = "display-setup"
version = "0.1.0"
description = ""
authors = [
    {name = "Jake Wood",email = "jake@ridgelineradio.org"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyserial (>=3.5,<4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
